requirements.txt
praw>=7.7.1
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
requests>=2.31.0
ollama>=0.1.7
chromadb>=0.4.15
sentence-transformers>=2.2.2
plotly>=5.17.0
streamlit>=1.28.1
click>=8.1.7
tqdm>=4.66.1
python-dateutil>=2.8.2 